+++ app/services/policies/create_policy.rb
@@ -10,7 +10,7 @@ module Policies
     def call
       Resource.transaction do
         resource = Resource.create(service: @service,
-                                   path: @json_body['path'],
+                                   pretty_path: @json_body['path'],
                                    resource_type: :local)
         create_access_policies(resource)
         create_user_managers(resource)